# replace RWMutex-map with skipmap