--- a/go.mod
+++ b/go.mod
@@ -9,5 +9,6 @@ require (
 	github.com/dustin/go-humanize v1.0.0
 	github.com/pkg/errors v0.9.1
 	github.com/stretchr/testify v1.4.0
+	github.com/zhangyunhao116/skipmap v0.1.1
 	golang.org/x/sys v0.0.0-20200930185726-fdedc70b468f
 )