Yesterday i saw an interesting article(https://dgraph.io/blog/post/introducing-ristretto-high-perf-go-cache/), i am faced with the same problem before, so i develop a high-performance map called `skipmap`(https://github.com/zhangyunhao116/skipmap).

Here is an alternative choice that replace `RWMutex-map` with `skipmap `.
```
name                old time/op    new time/op    delta
SketchIncrement-16    8.22ns ± 0%    8.24ns ± 1%      ~     (p=0.118 n=9+9)
SketchEstimate-16     8.35ns ± 1%    8.40ns ± 0%    +0.55%  (p=0.043 n=10+8)
StoreGet-16           34.6ns ± 0%    14.4ns ± 6%   -58.52%  (p=0.000 n=9+10)
StoreSet-16            396ns ± 2%     545ns ± 2%   +37.66%  (p=0.000 n=10+10)
StoreUpdate-16         410ns ± 2%     547ns ± 1%   +33.48%  (p=0.000 n=10+8)

name                old speed      new speed      delta
SketchIncrement-16   122MB/s ± 0%   121MB/s ± 1%      ~     (p=0.130 n=9+9)
SketchEstimate-16    120MB/s ± 1%   119MB/s ± 0%    -0.55%  (p=0.043 n=10+8)
StoreGet-16         28.9MB/s ± 0%  69.7MB/s ± 7%  +141.33%  (p=0.000 n=9+10)
StoreSet-16         2.53MB/s ± 2%  1.83MB/s ± 2%   -27.46%  (p=0.000 n=10+10)
StoreUpdate-16      2.44MB/s ± 1%  1.83MB/s ± 1%   -25.24%  (p=0.000 n=9+8)
```

The benchmark contains 100% Get or 100%Set, but usually, we often perform 50%Get50%Set, 30%Set70%Get, etc. the `skipmap` is a good choice for these situations, we can see the benchmark in https://github.com/zhangyunhao116/skipmap#benchmark

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/dgraph-io/ristretto/250)
<!-- Reviewable:end -->